Changed useState and useEffect location from Hero component to index
