+++ src/pages/index.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import { graphql, useStaticQuery } from 'gatsby';
 import Helmet from 'react-helmet';
 import { ThemeProvider } from 'styled-components';
@@ -13,6 +13,23 @@ import Portfolio from '../components/Portfolio';
 import Contact from '../components/Contact';
 
 const IndexPage = () => {
+  const [isMobile, setMobile] = useState(null);
+
+  /* eslint-disable */
+  const changeMobile = () => {
+    // FOR TESTS
+    console.log(window.innerWidth / 16);
+    // check if passed media query matches with window dimensions
+    window.matchMedia('(max-width: 37.5em)').matches ? setMobile(true) : setMobile(false);
+  };
+  /* eslint-enable */
+
+  useEffect(() => {
+    changeMobile();
+    window.addEventListener('resize', changeMobile);
+    return () => window.removeEventListener('resize', changeMobile);
+  }, []);
+
   const data = useStaticQuery(graphql`
     query {
       site {
@@ -39,8 +56,8 @@ const IndexPage = () => {
           <meta property="og:site_name" content={data.site.siteMetadata.title} />
           <meta property="og:description" content={data.site.siteMetadata.description} />
         </Helmet>
-        <Hero />
-        <About />
+        <Hero isMobile={isMobile} />
+        <About isMobile={isMobile} />
         <Portfolio />
         <Contact />
       </div>
